requirements.txt
spherical-dev[dev]
